requirements.txt
Django==2.2.12
facebook-sdk==3.1.0
django-bootstrap4==1.1.1
django-widget-tweaks==1.4.8
djongo==1.3.1
dnspython==1.16.0
Pillow==7.1.2